add CVSS metrics in search JSON output
